+++ packages/fr_mvvm_env/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0 
+- feat FrEnvDropdownView; deprecated EnvDropdownView
+
 ## 1.0.0 2025-8-18
 * rename from `fr_env_mvvm` package
+* feat 
+  - EnvModel
+  - IEnvViewModel, FrEnvViewModel
+  - EnvDropdownView
 * adapt Flowr 2.0.1